Re org files. closes #26 (#27)
* move field_name and params into __init__ #24

* change field_name and params in example.

* save

* made insert more efficient. execute 2 fetches to get all data from ib.

* closes #26. move IB to connector specific file

* add custom vs code spell settings.

* update rest of examples with connector specific files
