+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "cSpell.words": [
+        "ECBOT",
+        "GLOBEX",
+        "NYMEX",
+        "groupby",
+        "insync"
+    ]
+}
