+++ examples/moving_average_crossover_stocks.py
@@ -1,5 +1,7 @@
 from loguru import logger
-from ta_scanner.data import load_and_cache, IbDataFetcher
+
+from ta_scanner.data.data import load_and_cache
+from ta_scanner.data.ib import IbDataFetcher
 from ta_scanner.indicators import IndicatorSmaCrossover, IndicatorParams
 from ta_scanner.signals import Signal
 from ta_scanner.filters import FilterCumsum, FilterOptions, FilterNames
