+++ ta_scanner/filters.py
@@ -20,15 +20,16 @@ class FilterException(Exception):
 
 
 class BaseFitler(metaclass=abc.ABCMeta):
-    def __init__(self):
-        pass
+    def __init__(self, field_name: str, params: Dict[FilterOptions, Any]):
+        self.field_name = field_name
+        self.params = params
 
     def ensure_required_filter_options(
         self, expected: List[FilterOptions], actual: Dict[FilterOptions, Any]
     ):
         for fo_key in expected:
             if fo_key not in actual:
-                raise FilterException(f"expected this key key = {fo_key}")
+                raise FilterException(f"expected key = {fo_key}")
 
     @abc.abstractmethod
     def apply(self, df, field_name, filter_options):
@@ -50,23 +51,15 @@ class FilterCumsum(BaseFitler):
     def log_entry(self, action, row):
         logger.debug(f"Action={action}. Ts={row.ts}. Close={row.close}")
 
-    def apply(
-        self,
-        df: pd.DataFrame,
-        field_name: str,
-        filter_options: Dict[FilterOptions, Any],
-        inverse: int = 1,
-    ):
-        self.ensure_required_filter_options(
-            self.required_filter_options, filter_options
-        )
+    def apply(self, df: pd.DataFrame, inverse: int = 1):
+        self.ensure_required_filter_options(self.required_filter_options, self.params)
 
-        query_signals = f"{field_name} != 0"
+        query_signals = f"{self.field_name} != 0"
 
-        threshold = filter_options[FilterOptions.threshold_intervals]
+        threshold = self.params[FilterOptions.threshold_intervals]
 
         for index, rs in df.query(query_signals).iterrows():
-            signal_direction = rs[field_name] * inverse
+            signal_direction = rs[self.field_name] * inverse
             self.log_entry(signal_direction, rs)
 
             for index_after in range(0, threshold):
@@ -82,12 +75,12 @@ class FilterCumsum(BaseFitler):
                 rx = df.iloc[df_index]
                 diff = (rx.close - rs.close) * signal_direction
 
-                if diff >= filter_options[FilterOptions.win_points]:
+                if diff >= self.params[FilterOptions.win_points]:
                     self.log_exit("Won", diff, df.loc[df_index])
                     df.at[df_index, self.name] = diff
                     break
 
-                if diff <= (filter_options[FilterOptions.loss_points] * -1.0):
+                if diff <= (self.params[FilterOptions.loss_points] * -1.0):
                     df.at[df_index, self.name] = diff
                     self.log_exit("Lost", diff, df.loc[df_index])
                     break
