+++ tests/test_filters.py
@@ -10,27 +10,32 @@ def gen_df_zeros(field_name="some_field_name"):
 
 
 def test_abstract_methods_present():
-    FilterCumsum()
+    field_name, params = "some_field_name", []
+    FilterCumsum(field_name=field_name, params=params)
 
 
 def test_required_filter_options():
     field_name = "indicator_name"
     df = gen_df_zeros(field_name)
-    filter_cumsum = FilterCumsum()
 
-    filter_options: Dict[FilterOptions, Any] = {
+    params: Dict[FilterOptions, Any] = {
         FilterOptions.win_points: 20.0,
-        # FilterOptions.loss_points: 10.0,
+        FilterOptions.loss_points: 10.0,
         # FilterOptions.threshold_intervals: 50,
     }
 
-    with pytest.raises(FilterException):
-        filter_cumsum.apply(df, field_name, filter_options)
+    filter_cumsum = FilterCumsum(field_name=field_name, params=params)
 
-    filter_options: Dict[FilterOptions, Any] = {
+    with pytest.raises(FilterException) as execinfo:
+        filter_cumsum.apply(df)
+        assert "FilterOptions.loss_points" in str(excinfo.value)
+
+    params: Dict[FilterOptions, Any] = {
         FilterOptions.win_points: 20.0,
         FilterOptions.loss_points: 10.0,
         FilterOptions.threshold_intervals: 50,
     }
 
-    filter_cumsum.apply(df, field_name, filter_options)
+    filter_cumsum = FilterCumsum(field_name=field_name, params=params)
+
+    filter_cumsum.apply(df)
