+++ tests/test_indicators.py
@@ -14,22 +14,23 @@ def gen_df_zeros(field_name="some_field_name"):
 
 
 def test_abstract_methods_present():
-    IndicatorSmaCrossover()
+    field_name, params = "field_name", []
+    IndicatorSmaCrossover(field_name=field_name, params=params)
 
 
 def test_ensure_required_filter_options():
-    sma_crossover = IndicatorSmaCrossover()
-
-    fake_df = gen_df_zeros()
-    fake_field_name = "fake_some_name"
+    field_name = "fake_some_name"
+    fake_df = gen_df_zeros(field_name)
 
     params = {
         IndicatorParams.fast_sma: 20,
         # IndicatorParams.slow_sma: 50, # intentionally missing param
     }
 
+    sma_crossover = IndicatorSmaCrossover(field_name=field_name, params=params)
+
     with pytest.raises(IndicatorException) as e:
-        sma_crossover.apply(fake_df, fake_field_name, params=params)
+        sma_crossover.apply(fake_df)
 
     expected_message = "IndicatorSmaCrossover requires key = IndicatorParams.slow_sma"
     assert expected_message == str(e.value)
